requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.5.0; python_version >= '3.7'
django-filter==21.1
django==4.0.4
djangorestframework-simplejwt==5.1.0
djangorestframework==3.13.1
pyjwt==2.3.0; python_version >= '3.6'
pytz==2022.1
sqlparse==0.4.2; python_version >= '3.5'
